server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
cryptography==41.0.7
Pillow==10.1.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
boto3==1.34.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0